(split)
Formula for the amount of data that had to be in the buffer
before it could be encoded was incorrect.  Old formula
(for the polyphase filterbank) was:

old formula:  286+576*(1+gfc->mode_gr) = 2014  (MPEG1)

Correct formula is:  512+framesize -32 = 1632

Best way to see this is to go through window_subband() in newmdct.c
It access samples x[32*i] through x[510+32*i]  for i=0..17.

So our old condition was too conservative.

However, the FFT still requires 1904 samples, so the
net change is just to lower the requirement from 2014 to 1904.


git-svn-id: https://svn.code.sf.net/p/lame/svn/trunk@3284 981c7233-bbb0-e711-95a9-001517a2e1a4